Remove incorrect use of 'is' in comparisons (#1264)
'is' and 'is not' check referential equality, not value equality.  As an
implementation detail, CPython caches strings so that _often_ something
like
    x = "defined"
    x is "defined"
will be True, but _not always_.  For example, if the string is
constructed differently, like
    x = "".join("defined")
    x is "defined"
is False (but `x == "defined"` is True).

Comparison to singleton literals (False, True and None) is fine and
preferred (their singleton nature guarantees that they will always be
referentially identical).